Database/requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
langchain
langchain-community
langchain-google-genai
